RestaurantMapper deleted. Use extension funtion in Extensions instead
